Arduino IDE code for standalone ESP8266 to update a twitter feed using thingspeak API
I had found code for updating a twitter feed using NodeMCU and LUA to
program the ESP8266 directly by user ok1cdj, but didn't find much help
with doing this in the Arduino IDE which supports programming the
ESP8266 directly. This code when flashed to the ESP8266 allows it to
connect to an access point and update a twitter feed with the account
linked at thingspeak.com. Enjoy!